gfx/Graphics: Move to own file
