+++ include/gfx/seadProjection.h
@@ -2,28 +2,13 @@
 #define SEAD_PROJECTION_H_
 
 #include <basis/seadTypes.h>
+#include <gfx/seadGraphics.h>
 #include <math/seadMatrix.h>
 #include <math/seadVector.h>
 #include <prim/seadRuntimeTypeInfo.h>
 
 namespace sead
 {
-namespace Graphics
-{
-enum DevicePosture
-{
-    cDevicePosture_Same = 0,
-    cDevicePosture_RotateRight = 1,
-    cDevicePosture_RotateLeft = 2,
-    cDevicePosture_RotateHalfAround = 3,
-    cDevicePosture_FlipX = 4,
-    cDevicePosture_FlipY = 5,
-    cDevicePosture_FlipXY = 3,
-    cDevicePosture_Invalid = 4,
-};
-
-}
-
 class Projection
 {
     SEAD_RTTI_BASE(Projection)
